FEATURE: Introduce a new releases plan NodeType
Usefull to display the rough release plan which was decided on the last sprint, see https://discuss.neos.io/t/release-cycle-and-versioning/928.

The data and all assets are loaded synchronously in the current NodeType, since it isn't used on most pages and we will re-implement the NodeTypes generally for the new site which is coming soon.