FEATURE: Provide Docker Compose setup for development
This is a work in progress for a Docker Compose setup
in order to develop neos.io. Some of the packages currently
used in the docker-compose.yml are not yet accessible for
everyone, but will be in the near future. They are those
images used in production for www.neos.io (apart from
Elasticsearch).