BUGFIX - Pagination
The pagination would not occur due to different query handling in paginate-count and marketplace query.
The reason was the "abandoned" field that was set to be excluded. In reality it is set anyways, but in abandoned packages is being set to 1 (string).
Removed exist query and added it to fusion so as to match the index. Now pagination shows up again